refactor: 불필요한 컴포넌트 삭제
